CHECK

+++ .shell/zcheckmd5.sh
@@ -1,8 +1,11 @@
 #!/bin/sh
 
+source /opt/config/mod/.shell/0.sh
+
 if [ -f /ZMOD ]; then
-    echo "Проверка системы из Klipper 12 не поддерживается. Используйте родной Klipper 11."
-    exit
+    DIR="/opt/config/mod/.shell/root"
+else
+    DIR="/opt/config/mod/.shell"
 fi
 
 restore_file()
@@ -19,46 +22,58 @@ restore_file()
 
 echo "Началась проверка. Она может занять много времени...."
 
-find /opt/PROGRAM/ -name md5sum.list | while read a;
-    do
-        b=$(pwd)
-        c=$(echo $a|sed 's/md5sum.list//')
-        echo "$c"
-        cd "$c"
-        if echo $c | grep -q control; then
-            touch Update
-        fi
-        md5sum -c md5sum.list 2>/dev/null | grep -v -e "OK$"
-        if echo $c | grep -q control; then
-            rm -f Update
-        fi
-        cd "$b"
-    done
+if ! [ -f /ZMOD ]; then
+    find /opt/PROGRAM/ -name md5sum.list | while read a;
+        do
+            b=$(pwd)
+            c=$(echo $a|sed 's/md5sum.list//')
+            echo "$c"
+            cd "$c"
+            if echo $c | grep -q control; then
+                touch Update
+            fi
+            md5sum -c md5sum.list 2>/dev/null | grep -v -e "OK$"
+            if echo $c | grep -q control; then
+                rm -f Update
+            fi
+            cd "$b"
+        done
+fi
 
 echo "/"
 cd /
 FF_VERSION="$(cat /root/version)"
 MIN_VERSION="3.1.3"
-if [ "${FF_VERSION//./}" -lt "${MIN_VERSION//./}" ]; then
-    sed '/\/nim\//d' /opt/config/mod/.shell/md5sum.list >/opt/config/mod/.shell/md5sum_nim.list
-    md5sum -c /opt/config/mod/.shell/md5sum_nim.list 2>/dev/null | grep -v -e "OK$" | tee /opt/config/mod/bad.list
-    rm -f /opt/config/mod/.shell/md5sum_nim.list
+if [ "${FF_VERSION//./}" -lt "${MIN_VERSION//./}" ] && ! [ -f /ZMOD ]; then
+    sed '/\/nim\//d' ${DIR}/md5sum.list >${DIR}/md5sum_nim.list
+    md5sum -c ${DIR}/md5sum_nim.list 2>/dev/null | grep -v -e "OK$" | tee /opt/config/mod/bad.list
+    rm -f ${DIR}/md5sum_nim.list
 else
-    md5sum -c /opt/config/mod/.shell/md5sum.list 2>/dev/null | grep -v -e "OK$" | tee /opt/config/mod/bad.list
+    md5sum -c ${DIR}/md5sum.list 2>/dev/null | grep -v -e "OK$" | tee /opt/config/mod/bad.list
 fi
 
 if [ "$1" == "restore" ]; then
-    cat /opt/config/mod/bad.list|grep ": FAILED$"|sed 's|: FAILED||' | sed 's|^./|/|' | while read a; do restore_file "$a"; done
+    if [ -f /ZMOD ]; then
+        echo "Переустановите zmod с флешки"
+    else
+        cat /opt/config/mod/bad.list|grep ": FAILED$"|sed 's|: FAILED||' | sed 's|^./|/|' | while read a; do restore_file "$a"; done
+    fi
 fi
 rm -f /opt/config/mod/bad.list
 
-if [ "${FF_VERSION//./}" -lt "${MIN_VERSION//./}" ]; then
-    sed '/\/nim\//d' /opt/config/mod/.shell/list.link >/opt/config/mod/.shell/md5sum_nim.list
-    chmod +x /opt/config/mod/.shell/md5sum_nim.list
-    /opt/config/mod/.shell/md5sum_nim.list 2>/dev/null
-    rm -f /opt/config/mod/.shell/md5sum_nim.list
+if [ "${FF_VERSION//./}" -lt "${MIN_VERSION//./}" ] && ! [ -f /ZMOD ]; then
+    sed '/\/nim\//d' ${DIR}/list.link >${DIR}/md5sum_nim.list
+    chmod +x ${DIR}/md5sum_nim.list
+    ${DIR}/md5sum_nim.list 2>/dev/null
+    rm -f ${DIR}/md5sum_nim.list
 else
-    /opt/config/mod/.shell/list.link 2>/dev/null
+    ${DIR}/list.link 2>/dev/null
 fi
 
+if ! [ -f /ZMOD ]; then
+    echo "Самопроверка zmod"
+    [ ${FF5X} -eq 0 ] && [ "$1" != "init" ] && umount ${UMOUNT_MOD}
+    chroot ${MOD} /opt/config/mod/.shell/zcheckmd5.sh
+    [ ${FF5X} -eq 0 ] && [ "$1" != "init" ] && mount --bind ${REMOUNT_MOD} ${UMOUNT_MOD}
+fi
 echo "Оригиналы файлов можно найти по ссылке https://github.com/ghzserg/zmod/tree/main/stock"
